BASIC lib clean up

+++ src/lib/basic/LELLIPSE.c
@@ -1,7 +1,7 @@
 
 #include <basic.h>
 
-void LELLIPSE(int x, int y, int w, int h, byte c)
+void LELLIPSE(int x, int y, int w, int h, int c)
 {
     LARC(x, y, w, h, 0, 256, c);
 }
